requirements.txt
annotated-types
anyio
certifi
click
distro
fastapi
greenlet
h11
httpcore
httpx
idna
jiter
openai
psycopg2-binary
pydantic
pydantic_core
PyPDF2
python-dotenv
python-multipart
sniffio
SQLAlchemy
starlette
tqdm
typing-inspection
typing_extensions
uvicorn
